+++ src/components/register.js
@@ -1,4 +1,4 @@
-import { registerUser } from '../lib/index';
+import { registerUser } from '../lib/credentials.js';
 
 function register(navigateTo) {
   const section = document.createElement('section');
@@ -22,7 +22,7 @@ function register(navigateTo) {
   buttonCreateAccount.classList.add('createAccount');
   buttonCreateAccount.addEventListener('click', () => {
     if (inputPassword.value === inputConfirmPassword.value) {
-      registerUser(inputName.value, inputUser.value, inputPassword.value);
+      registerUser(inputUser.value, inputPassword.value);
       navigateTo('/timeline');
     }
   });
