--- src/lib/firebase/initializeFirebase.js
@@ -1,16 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from 'firebase/app';
-import { getAuth,createUserWithEmailAndPassword } from "firebase/auth";
-import { firebaseConfig } from './credentials';
-
-
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const auth = getAuth(app);
-
-export{auth,app,createUserWithEmailAndPassword}
