--- src/lib/index.js
@@ -1,10 +1,0 @@
-// aqui exportaras las funciones que necesites
-import {createUserWithEmailAndPassword} from './firebase/initializeFirebase'
-
-export const registerUser = () => {
-  try{
-    createUserWithEmailAndPassword (auth, email, password)
-  } catch (error){
-    throw error.message;
-  }
-};
