+++ src/lib/localStore.js
@@ -1,0 +1,4 @@
+// aqui exportaras las funciones que necesites
+
+
+
